XIVY-14357 automatically fill name of add dialog
the name field fill now be automatically filled with 'NewVariable'

if that variable name already exists on the initially selected namespace, an index will be added to the end. e.g. 'NewVariable2', 'NewVariable3', etc.

added unittests for finding the name to fill in and improved playwright test to check for it to be present when opening the add dialog.